RateRepository index type bugfix

+++ src/main/java/com/exmachina/sam/currency/repositories/RateRepository.java
@@ -5,10 +5,11 @@ import com.exmachina.sam.currency.projection.RateSourceProjection;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface RateRepository extends JpaRepository<Rate, Integer> {
+public interface RateRepository extends JpaRepository<Rate, BigInteger> {
     Rate findBySourceAndDestination(String sourceCurrency, String destinationCurrency);
     List<Rate> findByDestination(String destinationCurrency);
     List<Rate> findBySource(String sourceCurrency);
